condense oauth imports
