+++ project/spotipy/__init__.py
@@ -4,7 +4,8 @@
 # from .oauth2 import *  # noqa
 # from .util import *  # noqa
 
-from spotipy import oauth
 from spotipy.errors import SpotifyException, SpotifyHttpError, \
     SpotifyStateError, SpotifyOAuthError
+
+from spotipy import oauth
 from spotipy.oauth import ClientCredentialsFlow, AuthorizationFlow, PKCEFlow
