+++ project/spotipy/oauth/auth/servers.py
@@ -15,7 +15,7 @@ from spotipy import errors
 # Templates used for generating responses
 # from the local server are stored in this
 # path.
-TEMPLATE_ROOT = pathlib.Path(__file__).parent[1] / "templates"
+TEMPLATE_ROOT = pathlib.Path(__file__).parents[1] / "templates"
 
 # Identify the appropriate usecase for
 # each template.
