+++ project/spotipy/oauth/cache/base.py
@@ -1,15 +1,15 @@
 import abc, typing
 
-from spotipy.oauth import utils
+from spotipy import oauth
 
 class SpotifyCacheHandler(typing.Protocol):
 
-    def save_token_data(self, token_data: utils.TokenData) -> None:
+    def save_token_data(self, token_data: oauth.TokenData) -> None:
         """
         Save the given token data to the cache.
         """
 
-    def find_token_data(self) -> utils.TokenData | None:
+    def find_token_data(self) -> oauth.TokenData | None:
         """
         Retrieve token data from the cache. If
         not found, return `None`
@@ -19,13 +19,13 @@ class SpotifyCacheHandler(typing.Protocol):
 class BaseCacheHandler(abc.ABC):
 
     @abc.abstractmethod
-    def save_token_data(self, token_data: utils.TokenData) -> None:
+    def save_token_data(self, token_data: oauth.TokenData) -> None:
         """
         Save the given token data to the cache.
         """
 
     @abc.abstractmethod
-    def find_token_data(self) -> utils.TokenData | None:
+    def find_token_data(self) -> oauth.TokenData | None:
         """
         Retrieve token data from the cache. If
         not found, return `None`
